refactor: 의존성 역전 제거
