--- src/main/java/chess/controller/request/Input.java
@@ -1,8 +1,0 @@
-package chess.controller.request;
-
-import java.util.List;
-
-public interface Input {
-
-    List<String> inputGameCommand();
-}
