--- src/main/java/chess/controller/resposne/Output.java
@@ -1,14 +1,0 @@
-package chess.controller.resposne;
-
-import java.util.List;
-
-public interface Output {
-
-    void printInitialMessage();
-
-    void printBoard(List<List<PieceResponse>> piecePosition);
-
-    void printStatus(StatusResponse statusResponse);
-
-    void printError(Exception e);
-}
