+++ src/main/java/chess/view/InputView.java
@@ -1,15 +1,13 @@
 package chess.view;
 
-import chess.controller.request.Input;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
 
-public class InputView implements Input {
+public class InputView {
 
     private static final Scanner scanner = new Scanner(System.in);
 
-    @Override
     public List<String> inputGameCommand() {
         return Arrays.asList(scanner.nextLine().split(" ", -1));
     }
