+++ src/test/java/chess/controller/resposne/ColorMapperTest.java
@@ -3,6 +3,7 @@ package chess.controller.resposne;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.piece.Color;
+import chess.view.resposne.ColorMapper;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.params.ParameterizedTest;
