+++ src/test/java/chess/controller/resposne/PieceMapperTest.java
@@ -3,6 +3,7 @@ package chess.controller.resposne;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import chess.domain.piece.PieceType;
+import chess.view.resposne.PieceMapper;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.params.ParameterizedTest;
